moved the isLoading and setIsLoading props down two levels to make the skeleton loaders for the venue cards
