+++ src/components/DataProvider/index.jsx
@@ -1,26 +1,28 @@
 import useFetch from "../../hooks/useFetch.jsx";
 import { useState, useEffect } from "react";
 // import { Outlet } from "react-router-dom";
-import { createContext } from "react";
+// import { createContext } from "react";
+import { DataContext } from "../../contexts/index.jsx";
 
-const page = 1;
-const limit = 10;
+// const page = 1;
+// const limit = 10;
 // ?page=${page}&limit=${limit}
 
 const apiVenuesUrl = import.meta.env.VITE_API_VENUES_URL;
 const sortedByRating = `${apiVenuesUrl}?sort=created&sortOrder=desc`;
 
-export const DataContext = createContext();
-
 export function DataProvider({ children }) {
-  const { data, isLoading, isError } = useFetch(`${sortedByRating}`);
+  const { data, isLoading, setIsLoading, isError } = useFetch(`${sortedByRating}`);
   const [venues, setVenues] = useState([]);
   const allVenuesArr = data.data;
   // const location = useLocation();
   // const page = location.pathname;
 
+  console.log("venues", venues);
+
   //fetched all venues from the API and filter out the ones with missing location data, so that the search results will be accurate. Would wish that this was already fixed by the api, so that i could fetch it by page ang with a page limit, and not fetch all.
   useEffect(() => {
+    setIsLoading(true);
     if (allVenuesArr && allVenuesArr.length >= 0) {
       const filteredOutMissingLocations = allVenuesArr.filter((venue) => {
         return venue.location.city && venue.location.country;
@@ -30,5 +32,5 @@ export function DataProvider({ children }) {
     }
   }, [data, allVenuesArr]);
 
-  return <DataContext.Provider value={{ venues, setVenues, isLoading, isError }}>{children}</DataContext.Provider>;
+  return <DataContext.Provider value={{ venues, setVenues, isLoading, setIsLoading, isError }}>{children}</DataContext.Provider>;
 }
